requirements/base.txt
pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.2  # https://github.com/un33k/python-slugify
Pillow==6.0.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat

# Django
# ------------------------------------------------------------------------------
django==2.2.2  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.3  # https://github.com/django-compressor/django-compressor
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.4  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client


#ps
mysqlclient==1.4.2.post1
django-taggit #https://github.com/jazzband/django-taggit
django-markdownx #https://github.com/neutronX/django-markdownx
sorl-thumbnail==12.5.0 #https://sorl-thumbnail.readthedocs.io/en/latest/installation.html



#xadmin 需要
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools==2.1
future==0.15.2
httplib2==0.9.2
six==1.10.0
xlwt
xlsxwriter
